Put info for Readme
# Friendly Collaboration Party

This repository is for planning a party to learn GitHub and more!

## About
This is some text about my project.
Parties are great right? Let me tell you how mine will be the very best.

## Team
So far my team is small - just 1 person, but I'm willing to work with you!
I'd really like for you to bring a few more people along to join in.

## Setup & Requirements
This project requires a computer and internet.